comment

+++ src/main/java/life/majiang/community/community/dto/QuestionDTO.java
@@ -3,14 +3,17 @@ package life.majiang.community.community.dto;
 import life.majiang.community.community.model.User;
 import lombok.Data;
 
+/**
+ * @author Administrator
+ */
 @Data
 public class QuestionDTO {
-    private Integer id;
+    private Long id;
     private String title;
     private String description;
     private Long gmtCreate;
     private Long gmtModified;
-    private Integer creator;
+    private Long creator;
     private Integer commentCount;
     private Integer viewCount;
     private Integer likeCount;
